requirements.txt
asgiref==3.8.1
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.3.2
Django==4.0.3
django-cors-headers==4.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
install==1.3.5
mccabe==0.6.1
mypy==0.941
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==1.7.1
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-requests==2.27.15
types-urllib3==1.26.11
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
